도서, 서재, 서재 관련 Dialog, 데이터 흐름 수정 및 레이어 구분 + concatadapter 제거

+++ BookChat/app/src/main/java/com/example/bookchat/utils/BookImgSizeManager.kt
@@ -40,8 +40,8 @@ object BookImgSizeManager {
     val bookImgHeightPx =
         (DEFAULT_BOOK_IMG_HEIGHT_PX * scaleRate).toInt()
 
-    fun getFlexBoxDummyItemCount(totalItemCount: Long) =
-        (flexBoxBookSpanSize - (totalItemCount % flexBoxBookSpanSize).toInt()) % flexBoxBookSpanSize
+    fun getFlexBoxDummyItemCount(totalItemCount: Int) =
+        (flexBoxBookSpanSize - (totalItemCount % flexBoxBookSpanSize)) % flexBoxBookSpanSize
 
     fun getPxFromDp(dp: Int): Int =
         (dp * (displayMetrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)).toInt()
